Use type float32 instead of float64 for storing hyper parameters.

+++ cpp/subprojects/boosting/src/mlrl/boosting/rule_evaluation/rule_evaluation_decomposable_partial_dynamic.cpp
@@ -24,9 +24,9 @@ namespace boosting {
 
             DenseScoreVector<PartialIndexVector> scoreVector_;
 
-            const float64 threshold_;
+            const float32 threshold_;
 
-            const float64 exponent_;
+            const float32 exponent_;
 
             const float32 l1RegularizationWeight_;
 
@@ -50,7 +50,7 @@ namespace boosting {
                                                      float32 exponent, float32 l1RegularizationWeight,
                                                      float32 l2RegularizationWeight)
                 : outputIndices_(outputIndices), indexVector_(outputIndices.getNumElements()),
-                  scoreVector_(indexVector_, true), threshold_(1.0 - threshold), exponent_(exponent),
+                  scoreVector_(indexVector_, true), threshold_(1.0f - threshold), exponent_(exponent),
                   l1RegularizationWeight_(l1RegularizationWeight), l2RegularizationWeight_(l2RegularizationWeight) {}
 
             const IScoreVector& calculateScores(StatisticVector& statisticVector) override {
